+++ urcu/static/urcu-bp.h
@@ -93,7 +93,7 @@ static inline void debug_yield_write(void)
 
 static inline void debug_yield_init(void)
 {
-	URCU_TLS(rand_yield) = time(NULL) ^ pthread_self();
+	URCU_TLS(rand_yield) = time(NULL) ^ (unsigned long) pthread_self();
 }
 #else
 static inline void debug_yield_read(void)
